Add newline for REPL output (#1926)
The current REPL library doesn't provide whitespace between the
expression and the output. With the REPL here, we have the ability to
color the prompt which helps. to visually group the expression with the
output. But with the PAC CLI, we don't have this ability and it all
bleeds together. This is particularly true for Power Fx where an
expression could cause recalc of other values which are rightly shown in
the output.

In short, we want to change this:

![image](https://github.com/microsoft/Power-Fx/assets/15149773/ff33efe5-552d-492e-a3c1-7a9151242ed5)

Into the slightly longer, but much easier to parse and read:

![image](https://github.com/microsoft/Power-Fx/assets/15149773/f0979d2c-1e71-4aab-b57f-b15f0ffda597)

We could have fixed this just in the PAC CLI, however I prefer this
whitespace here too and consistency is good. This also ensures that
table formatting works correctly with added newlines. It also mirrors
the tests cases.