Send back empty control tokens (#1925)
We previously anticipated wrongly that we do not need to send back empty
control tokens. We actually need to do send back empty control tokens
and that's the only way we know we went from having atleast one control
token in a formula to zero control tokens in the formula. This is a
small change that we need to do