Deprecate Coercion to Primary Output properties (#1968)
This adds a Feature to allow us to disable coercion to Primary Output
Properties.
This is a feature that only affected Canvas apps, as it depends on the
Control type, which does not exist in non-canvas hosts. Unfortunately,
to accomplish this, we need to pass `Features` throughout the DType tree
so that CoercesTo can have access to this flag.