Fix TestWithClockService test (#2026)
Fixes #2025. The case `TimeZoneOffset()` in the TestWithClockService test is failing: it assumes the Pacific Standard Time, and expects the result to be 420 (UTC-07:00). However, we are now out of the daylight saving time, so the current value is now 480. This test will now validate against the current offset for that time zone.