Implementation of SortByColumns function (#2011)
The function is defined in the core library, but didn't have an implementation yet. It does now.

This is a special function which takes columns names as arguments. In all other functions (such as [Add/Drop/Rename/Show]Columns) the column names must be literal strings (or identifiers, if SupportColumnsAsIdentifiers is enabled), but SortByColumns in Power Apps allows for arbitrary string-valued expressions to be used as column names. For Power Fx 1.0 we are going to block it for now, and add support for that later if we get feedback from the community.