Remove resources from translated file that don't exist in en-US (#2104)
During development we add resources from the en-US file, and as they are translated they are added to the resources for other languages. But when  we remove resources from en-US, they remain on the satellite resource files. This cleans them up for now (although we should have it automated to prevent this problem from happening again).