Refactor NL2Fx Handling in Language Server to support more dynamic NL Handler Creation based on IPowerFxScope (#2134)
This pr performs necessary wiring to enable creating NL Handler for
Nl2Fx and FX2NL based on the current IPowerFxScope language server is
operating on. This is necessary to support Nl2FX scenarios in Power Apps
and the current approach which uses Nl2FxImplementation property is not
a thread safe. With the plan to perform Nl2Fx operations in parallel and
given that we only create 1 instance of Language Server SDK, two
concurrent Nl2FX tasks can overwrite Nl2FxImplementation unexpectedly.
The new approach in this pr scopes the creation of handler to every
Nl2FX task. It allows the creation of handler based on the scope itself
which has "host" specific entities which are required to create a more
meaningful NlHandler. For Power Apps, "host" contains document, control
and property which are critical to context building for Nl2Fx model
inside NL Handler.