Update Void accept behavior and take advantage in Repl (#2096)
Several Power Fx functions have no real return value. To date, Canvas
apps have returned Boolean values that aren't seen and likely this
behavior is not well known. With the Repl, these values are being
printed as part of the stream and it can be confusing. to see `Set( x, 3
)` and then `true`. Similarly, it is seems odd in a strongly typed
language to return a value that really has no meaning (the `true` from
Set) which could be a coding error, the kind of thing that strong typing
should catch.

Changes to be made:
1. Void will accept all types, including Void. In Power Apps, behavior
properties would be of type Void, and they work with any formula,
dropping the computed value.
2. Void will not be accepted by any other types (except Void). Trying to
use the output from Set will result in an error.
3. The following functions will return Void: Set, Clear, Remove, Notify
(Repl), Help (Repl), Exit (Repl). Future functions that have no return
value will return Void.

There is a tiny loss of functionality with this change. The result from
Remove can no longer be put in a variable. Which normally doesn't matter
as Remove returns no value. But, if Remove encounters an error, that
error can no longer be put in a variable. The error will still be
reported and can be caught and processed with IfError and App.OnError in
Power Apps. With an IfError, the error record can be extracted and that
can be put in a variable. The team discussed this and we do not believe
this is a common pattern, if used at all.

With this change, it is also not possible to use Boolean logic to short
circuit the use of these functions. For example, it used to be possible
to `Back() && Clear( table )` and the Clear would only be executed if
the Back() returned true. We don't think the intent here is very clear,
especially for a non-pro developer, and we'd prefer it be written as
`If( Back(), Clear( table )`. Usage on Canvas apps is very low for this
pattern.