Fix handling of GMT time zone (#2175)
There were some parts of the code that were mistakenly treating the GMT (i.e., London) time zone as UTC. Those are the same for part of the year, but not for all of it, which was causing some issues with parsing and adding date values. This change addresses it.