Fix #2241. (#2339)
Fix #2241 - adds a TryGetType to lookup a variable by name (very useful
when it doesn't have a slot) and return its type.

This is particularly interesting for enums/optionsets - which don't have
slots (since they don't have per-eval storage). But
beware , #2342 - some of the FormulaTypes for enums/optionsets are not
correct.