Update Mod and RandBetween type handling (#2355)
From a type perspective, Mod should act like / (division). The arguments
should be symmetric with a Decimal result only if both arguments are
Decimal. This is not the case, there is a bug here where the first
argument is dictating the result of the function alone. The same holds
for RandBetween which has the same problem. This PR fixes both functions
and adds many tests.

Here, the second result should be 1 with no decimal as the calculation
should be done in Float.
```
>> Mod( 3.0000000000000000000003, 2.0000000000000000000002 )
1.0000000000000000000001

>> Mod( 3.0000000000000000000003, Float(2.0000000000000000000002) )
1.0000000000000000000003

>> Mod( Float(3.0000000000000000000003), 2.0000000000000000000002 )
1

>> Mod( Float(3.0000000000000000000003), Float(2.0000000000000000000002) )
1
```

Likewise, RandBetween has the wrong answer for the second case, where
the second argument is coerced to a decimal (while losing digits).
```
>> 
>> RandBetween( 12345678912345678912345678, 12345678912345678912345678 )
12345678912345678912345678

>> RandBetween( 12345678912345678912345678, Float(12345678912345678912345678) )
12345678912345697471862876

>> RandBetween( Float(12345678912345678912345678), 12345678912345678912345678 )
1.2345678912345679E+25

>> RandBetween( Float(12345678912345678912345678), Float(12345678912345678912345678) )
1.2345678912345679E+25
```