Added another property on interface IExternalNamedFormula to detect reference to views in NF expressions. (#2522)
Currently the logic to detect use of dataverse views in an expression
doesn't work well when expression is abstracted under named formula.
For example,
T = Filter(Monthses, 'Monthses (Views)'.MyView);

Label1.Text = CountRows(T);

CountRows(T) is not detected currently as non-delegated expression as a
result. This change fixes that. Unit test will be added on PAClient side
for this where all of the delegation tests currently are.