Mark UDFs and NFs Async If there's a reference to cross screen control (#2620)
Marks UDF and Nf async if they refer to cross screen control and a new
binding config to do that is enabled. There's no definite way to say
where binder is binding nf or udf rules so config would be used to
communicate the marking of top parse node as async upon countering those
references. We don't need to mark those async if all they refer to is
component definition but there's no definite way to also check that.
ExternalTemplate.IsComponent returns true for component defs and
instances but we need to check only for the former. However that
shouldn't matter as code gen to lazy load component defs would be no op
since component defs are loaded upfront.