Fix formatting and structural printing for type literals (#2719)
Formatting , structural printing and cloning were broken for type
literals. This PR fixes these issues and adds tests to ensure the
changes work as expected.

Formatting of user definitions is still not perfect and has a bug where
it would fail if we had definitions of same kind with same name. That is
not in the scope of this PR and shall be fixed as a follow up