Fixes enum store when SymbolTable is replaced while setting up config (#2729)
This PR fixes the issue where inbuilt enums are no longer part of the
config if the config's symbol table is replaced after initialization.

After initializing the config if SymbolTable was replaced, we were
loosing the enums attached to the previous default symbol table. This PR
will adds a default symbol table(seprate from existing table) to handle
that.