Add UseDefaultBodyNameForSinglePropertyObject (#2815)
In Power Apps, when a body parameter is used it's flattened and we
create one paramaeter for each
body object property. With that logic each parameter name will be the
object property name.

When UseDefaultBodyNameForSinglePropertyObject is set, we will use the
real body name specified
in the swagger instead of the property name of the object, provided
there is only one property.