Adds Count Capabilites. (#2826)
This pull request introduces several changes to the PowerFx library to
support counting capabilities in table delegation. The changes include
adding a new method to count rows asynchronously, updating interfaces,
and introducing a new class for count capabilities.

Enhancements to counting capabilities:

*
[`src/libraries/Microsoft.PowerFx.Connectors/Public/CdpTableValue.cs`](diffhunk://#diff-993c4c638e07f59077c442100bdb8ae7ae8bb5d6a56cc003d71468556198f930R69-R73):
Added the `GetCountAsync` method to count rows asynchronously.
*
[`src/libraries/Microsoft.PowerFx.Core/Entities/External/TableDelegationInfo.cs`](diffhunk://#diff-3262aeee6122a5873eb3c7866b3db562c90530f04d665f6186b26f32328abcf9R29-R31):
Introduced the `CountCapabilities` class and added the
`CountCapabilities` property to `TableDelegationInfo`.
[[1]](diffhunk://#diff-3262aeee6122a5873eb3c7866b3db562c90530f04d665f6186b26f32328abcf9R29-R31)
[[2]](diffhunk://#diff-3262aeee6122a5873eb3c7866b3db562c90530f04d665f6186b26f32328abcf9R305-R327)
*
[`src/libraries/Microsoft.PowerFx.Core/Public/Values/TableValue.cs`](diffhunk://#diff-acf9bad0ca58398efcb3fd3f6804a99623c3719c070b4f2d35f1775d040458c2R31-R39):
Updated the `IDelegatableTableValue` interface to include the
`GetCountAsync` method.