Pass Feature Name as Telemetry for NL2Fx (#2848)
Currently for the telemetry object in NL2Fx, we only pass one unified
identifier for all NL2Fx features which results in data not being
classified at the feature level. We must change that, for better
telemetry information.

This is a four-part change. In these repo, Power-Fx (GitHub), I need to
write a variable in CustomNL2Fx, to pass to PPUX.