Fix DType.Add overloads (#2856)
`DType.Add` overload was broken and it did not work correctly for
LazyTypes leading to errors and NullReferenceExceptions in some cases.

this bug came to light when investigating a customer reported incident
for new analysis.

This PR fixes the `DType.Add` overload and adds tests to ensure the fix
works as expected.