Seal system records and tables, ensuring we can add fields in the future. (#2212)
Records and tables can now be sealed, ensuring that we won't break
formulas if we decide to add something to them at a later date.

This means that the User record, IfError's FirstError/AllErrors
record/table, and select other "system" records and tables that we
select in the future:
- will be blocked from being used in `AddColumns` and `RenameColumns`.
As this may result in a field being introduced by the maker that would
collide with something we may want to add later.
- can't participate in a type union with other records. `If( true,
{a:1}, User )` will result in a Void type.

`ShowColumns` is the easy workaround that removes the block, as it has
the maker specifically select the columns that they want to use and
newly introduced columns will not impact it. Note that `DropColumns`
doesn't have the same impact as newly added columns would remain.

This change has no impact on Canvas apps, as it requires V1 compat to be
enabled. While Canvas doesn't have the User record, it does have the
User function, for which we should mark as sealed once we enable the V1
flag there.