Adds capabilities for top level aggregation. (#2873)
This pull request introduces new capabilities for handling top-level
aggregations in the `Microsoft.PowerFx.Core` library. The changes
primarily focus on adding support for top-level aggregation capabilities
and updating related classes and enums.

### Enhancements to Aggregation Capabilities:

*
[`src/libraries/Microsoft.PowerFx.Core/Entities/External/TableDelegationInfo.cs`](diffhunk://#diff-3262aeee6122a5873eb3c7866b3db562c90530f04d665f6186b26f32328abcf9R32-R34):
Added a new property `TopLevelAggregationCapabilities` to the
`TableDelegationInfo` class and marked it as obsolete for preview
purposes.
*
[`src/libraries/Microsoft.PowerFx.Core/Entities/External/TableDelegationInfo.cs`](diffhunk://#diff-3262aeee6122a5873eb3c7866b3db562c90530f04d665f6186b26f32328abcf9R351-R362):
Introduced a new class `TopLevelAggregationCapabilities` with a method
`IsTopLevelAggregationSupported` to determine support for top-level
aggregation on a column.
*
[`src/libraries/Microsoft.PowerFx.Core/Public/Values/DelegationParameters.cs`](diffhunk://#diff-7d0c7991f2b0cafe204ca1ce9c4aafaa8424ef742874060d7dfb322de4a8c457R82-R84):
Added a new enum value `ApplyTopLevelAggregation` to
`DelegationParameterFeatures` to represent the ability to apply
top-level aggregations.