Update for REPL multilineProcessor (#2891)
Small bug with multiline detection in the REPL. The processor would not
think `$"{"hi"}"` is complete due to an off-by-one string position
accounting error. Tests didn't find this because I was generous with
spaces in the test cases for readability, which also meant the space
would absorb the extra 1 and the test would pass.