Update Match.Email (#2886)
Updates to regular expression support:
- Updated definition for `Match.Email` and `Match.Hyphen`. These new
definitions will have no impact on Canvas (`DefaultEnums` did not
change), but the new values are available to Canavas to use in the
future under an opt-in.
- Removed largely unnecessary setup `SupportColumnNamesAsIdentifiers` so
that Canvas will run these tests too. Some minor tweaks to replace
`DropColumns` and `AddColumns` with other forms that aren't sensitive to
this setting. Results should not have materially changed.
- Remove obsolete preview warning from `EnableRegExFunctions`. These
functions are now stable, GA, and ready for production.