Fixes issue with OptionSet Naming. (#2898)
This pull request includes updates to improve type handling in
connectors and expand test coverage by adding new fields and scenarios.
The key changes include adjustments to how connector types are
determined, updates to test cases to include new fields, and
modifications to test data to support the expanded scenarios.

### Connector Type Handling:
* Updated the `GetConnectorType` method in `OpenApiExtensions.cs` to use
the name of the `openApiParameter` instead of the hardcoded string
"Array" when creating a `SwaggerParameter`. This ensures consistency and
better alignment with the input parameter.

### Test Case Enhancements:
* Modified the `DVDynamicReturnType` test in
`PowerPlatformConnectorTests.cs` to adjust the expected structure of the
return type, reflecting updates to the connector schema.
* Updated the `SF_CdpTabular_GetTables` and `SF_CdpTabular` tests in
`PowerPlatformTabularTests.cs` to include new fields `picklist1` and
`picklist2` in the expected output, improving coverage for scenarios
involving picklist arrays.
[[1]](diffhunk://#diff-7b1be2cff9606912fdb9b09ff75af4e880829323c9e42a7252a1f93feb1ded67L935-R935)
[[2]](diffhunk://#diff-7b1be2cff9606912fdb9b09ff75af4e880829323c9e42a7252a1f93feb1ded67L1102-R1102)

### Test Data Updates:
* Added `picklist1` and `picklist2` fields to the `SF GetSchema.json`
test data. These fields are arrays of strings with predefined
enumerations, supporting new test scenarios for picklist handling.