Converts void to error for non sideeffecting expressions. (#2917)
Fixes https://github.com/microsoft/Power-Fx/issues/2221
This pull request introduces functionality to detect and handle void
expressions in non-side-effecting contexts by converting them into error
expressions. It includes updates to the core binding logic, feature
configurations, and associated tests.

### Core Binding Enhancements:
* Added a new `VoidToErrorTexlVisitor` class in
`src/libraries/Microsoft.PowerFx.Core/Binding/VoidToErrorTexlVisitor.cs`
to traverse expressions and ensure void expressions are flagged as
errors.
* Updated the `Run` method in `TexlBinding` within
`src/libraries/Microsoft.PowerFx.Core/Binding/Binder.cs` to invoke the
`VoidToErrorTexlVisitor` when side effects are disallowed.

### Test Enhancements:
* Added test cases in
`src/tests/Microsoft.PowerFx.Core.Tests.Shared/ExpressionTestCases/VoidToError.txt`
to validate the behavior of void-to-error conversion across various
expressions and scenarios.
* Added test cases in
`src/tests/Microsoft.PowerFx.Interepter.Tests.Shared/MutationScripts/Void.txt`
to validate the Void behavior when side effects are enabled, largely not
leading to an error.

### Breaking Language Change:
The behavior of `IsError` and `IfError` are being modified with respect
to Void, _only when side effects are disabled_. Since use of Void in
these functions with side effects disabled would not have been useful,
it would always treated it as a non-error, this change is not expected
to impact existing customers.

The reason we turn a blind eye to Void results in side effect enabled
scenarios, such as Canvas app Button.OnSelect, is so that makers can
write control flow with the If function. For example, in this case where
Orders {Value: Number} and BadSessions {Value: GUID} have incompatible
schemas:
```
If( allGood, 
    Collect( Orders, {Value:Int(TextInput1.Text)} ), 
    Collect( BadSessions, {Value:SessionGUID} ) 
)
```

Makers should be able to wrap that whole formula with `IfError` or
`IsError` to detect runtime problems and take appropriate actions. For
example, the value of TextInput1.Text in this example may not be a
parseable number.
 
But without side effects, the `If` function isn't about control flow,
the formula is expected to produce a single, non-Void return value.

---------

Co-authored-by: Greg Lindhorst <gregli@microsoft.com>